Add GPLv3 License
I would like to fork and modify this project for use in a public repo, however due to the lack of a license I’m not able to do so. 

I’ve only suggested GPLv3 since you’re using it for the CORS plugin, but any compatible free software license would be fine and I can update this PR if some other license would be preferable.